+++ hubdle-gradle-plugin/test/resources/config/documentation/changelog/1/CHANGELOG.md
@@ -14,6 +14,10 @@
 
 ### Updated
 
+- `com.gradle.enterprise:com.gradle.enterprise.gradle.plugin -> com.gradle:com.gradle.develocity-gradle-plugin`
+
+### Updated
+
 - `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.14`
 
 ## [0.1.0-rc.14] - 2021-11-25